Migrate @turf/great-circle to TypeScript
