+++ packages/turf-great-circle/package.json
@@ -66,11 +66,13 @@
     "tape": "^5.9.0",
     "tsup": "^8.4.0",
     "tsx": "^4.19.4",
+    "typescript": "^5.8.3",
     "write-json-file": "^6.0.0"
   },
   "dependencies": {
     "@turf/helpers": "workspace:*",
     "@turf/invariant": "workspace:*",
-    "@types/geojson": "^7946.0.10"
+    "@types/geojson": "^7946.0.10",
+    "tslib": "^2.8.1"
   }
 }
